renamed a bit more. fragmented into now three pieces from two - a base animation capability, a tts capability and then a top level wrapper (untested) that allows outsiders to drive the talking head in the same way as before hopefully - should have the same interface as before as far as outsiders are concerned. also removed the thrown error when no google tts audio supplied. also added an ability to pass an audio blob that is not an array buffer - since that is not transportable over the net easily - but rather you can pass a b64 encoded audio string which is then decoded insitu
