+++ modules/anim-moods.mjs
